+++ client/infrastructure/http/ExerciseApi.ts
@@ -1,12 +1,13 @@
 import { IExerciseApi } from '@/client/application/api/IExerciseApi';
+import { API_ROUTES } from '@/shared/config/routes';
 import { ExerciseResponseDto } from '@/shared/dtos';
 
 export class ExerciseApi implements IExerciseApi {
 	async getMeta(): Promise<{ lastUpdatedAt: string | null }> {
-		return fetch('/api/exercises/meta').then(r => r.json());
+		return fetch(API_ROUTES.exercises.meta).then(r => r.json());
 	}
 
 	async getAll(): Promise<ExerciseResponseDto[]> {
-		return fetch('/api/exercises/all').then(r => r.json());
+		return fetch(API_ROUTES.exercises.all).then(r => r.json());
 	}
 }