+++ presentation/web/components/Footer.tsx
@@ -1,0 +1,11 @@
+const Footer = () => {
+	return (
+		<footer className=''>
+			<div className='text-muted-foreground mx-auto flex h-16 max-w-5xl items-center justify-center px-6 text-xs'>
+				Built by Maksim Sc
+			</div>
+		</footer>
+	);
+};
+
+export default Footer;